Read the connection information from a configuration file (versaplexd.ini) rather than using hardcoded constants in the source file.  Also don't bother caching the results of this lookup, as it's cheap (the INI file is in an in-memory hash anyway).
The config file is an INI file that has two sections: [User Map] and
[Connections].  [Connections] contains a list of arbitrary connection
identifiers as keys, with the value being a connection string.  For now,
the only connection strings recognized are of the form "mssql:Standard
MS SQL 2005 Connection String".  

[User Map] contains a list of user names as keys (taken from the Unix
passwd file for now), each one having the value of a key in
[Connections] that will be used as the connection string when that
user connects.

The ini file parsing routines live in wvdotnet, so we now have to refer
to wv.dll when building versaplexd.


git-svn-id: http://versaplex.googlecode.com/svn/trunk@284 5b722feb-fb3a-0410-8233-6f4790b74332