vxodbc: better connect error reporting.
In the SQLDriverConnect test, actually bounce a Test message to versaplexd
to ensure end-to-end connectivity is working *and* we're authenticating okay.
Report separate errors for "failed to connect" and "failed to get a reply."

Also fix Makefile to build correctly with newer mingw32-gcc (3.4.4 on my
machine) and output a vxodbc.map file in case I need to look at it.



git-svn-id: http://versaplex.googlecode.com/svn/trunk@296 5b722feb-fb3a-0410-8233-6f4790b74332