versaplexd: better error handling/reporting.
Add a new "Test" method that basically runs "select 1" on the database and
throws away the answer; call this to see if you can sucessfully connect to
the database you wanted.  Fix a problem with queries that aren't LIST TABLES
and whole length is less than 13 chars.  Use a default fallback username of
"*" if GetUnixUserName fails (which it always will on a Windows-based dbus
or a non-wvdbusd dbus).  Return the actual connection error string in the
dbus error message when connection fails for any reason.



git-svn-id: http://versaplex.googlecode.com/svn/trunk@297 5b722feb-fb3a-0410-8233-6f4790b74332