versaplexd can now run its own WvDBusServer in a thread.
We do this by creating wvdbusd.so, a little link library that pulls in
C++ WvStreams, including libwvdbus, and makes it accessible using a simple
C API.  Then we call into that .so using .net P/Invoke stuff.

Use the new "-l" option to versaplexd (one or more times) if you want it to
run a DBus daemon.