Move the VxColumnInfo class to its own file.
This means the unit tests can refer to the real version, instead of
maintaining their own copy-and-pasted version.  Also, since the two
implementations had (naturally) diverged, the class layout no longer
exactly matches the DBus wire format, so we have to read them in by hand
in the unit tests.