Explicitly write padding bytes, instead of moving the stream position.
Normally this isn't an issue, but if you have an empty array at the end
of a message, and the array elements require 8-byte padding, then DBus
requires 8-byte padding after the array size, regardless of the lack of
elements in the array.  Simply changing the stream position means that
those bytes will not actually be written when the message is sent.