Make PutSchema delegate reading and writing to VxSchemaElement.
Instead of reading and writing the DBus message ourselves when calling
or implementing PutSchema, just send and receive a VxSchemaElement.
This sends just about the same information, but reduces code duplication
and should be easier to maintain.

Also standardize on "type, name, text" order for specifying a schema
element everywhere; this is the format used for PutSchema, and really
makes more sense.  This changes the order of the VxSchemaElement
constructor and DBus wire format.