Support updating schemas on disk, by making Put act the same as the db backend.
In particular, if an element being Put has no text, delete it.  This
makes the results of VxSchema.GetDiffElements usable for updating only
the changed elements of an on-disk schema.

Also always return a VxSchemaErrors object from VxDiskSchema's Put, as
it matches the behaviour of the DBus and DB backends, and makes life
easier in the unit tests.