Add new ExecuteToWvAutoReader extension method.
This method combines SqlCommand.ExecuteReader with the existing
ToWvAutoReader extension method.  This ensures that the DataReader
created by ExecuteReader is cleaned up in a timely manner, which makes
WvDbi much better behaved with MSSQL servers.  MSSQL tends to get cranky
when more than one DataReader exists for a particular connection.