Throw an assertion if we have too many open connections.
MSSQL gets mad if it's got more than 100 open connections, and will hang when
opening any new ones.  Getting an assertion is much more helpful in this
situation, since (a) you get a clue about what's going on and (b) your
server process doesn't hang.

Give WvDbi a quota of 50 open connections; you can increase it if you
really need to, but that's probably not wise.