Create exactly one WvDbi per request, making sure to clean it up.
Instead of making VxDbSchema create WvDbis in its constructor, create one
when we receive a request, to be used throughout the entire request.  And
then be sure to clean it up when we're done.

As a corollary, VxDbSchema takes a WvDbi object in its constructor, rather
than the connection string whose main goal was to make a WvDbi (which led
to object lifetime issues).  The other use of the connection string, making
a direct connection to the database when WvDbi wouldn't serve, is now done
by borrowing WvDbi's connection temporarily.