Make DropSchema take a list of keys to drop, rather than a single element.
It was slow to drop a lot of elements, and further it didn't really match
the other Schemamatic API to only work on one element at a time.  So
now DropSchema takes a list of keys (dealing with separated types and
names was also a pain) and returns a set of VxSchemaErrors for any that
failed.