Replaced WvAutoCast[] with WvSqlRow.
There's no reason, most of the time, to create large arrays of
WvAutoCasts on SQL requests.  Also, we want to get the schema
information for a particular Query.  WvAutoCast[] doesn't have
this information.  What to do?

Apparently, create a wrapper object around object[] with an
overloaded this[] operator, which implements
IEnumerable<WvAutoCast> and a DataTable member!  Yay!