Move the code that parsed the database checksum strings into its own function.
VxDiskSchema was doing a lot of work in the middle of a busy function
just parsing a space-separated string of hex numbers.  Move that code to
the new static ParseSumString function in VxSchemaChecksum, since it
reverses the effects of GetSumString().

Also add a new VxSchemaChecksum constructor to take a pre-supplied
sequence of sums, such as provided by ParseSumString.