Use out parameters in VxDiskSchema.ReadSchemaFile.
Fiddling with the innards of parameters as a way to return information from
a function is a bit iffy.  Out parameters are more honest here, and cut
down on the creation of useless empty objects.  They also mean we can make
VxSchemaElement immutable.

Also make ReadSchemaFile private, as its API is a bit too hairy to expose
to the outside world.