Make VxSchemaChecksum immutable, then sensibly override Equals and GetHashCode.
We don't often change the contents of a VxSchemaChecksum, but we do have
to compare them for equality.  So it makes sense to officially make the
class immutable (so that it'll still work in hashtables), and override
Equals instead of having a silly "IsEqual" member function.  If you
override Equals though, you also have to override GetHashCode (again
so hashtables don't break).

Also move the unit tests for Equals, GetHashCode, and ParseSumString to
their own file, they don't depend on any of the other classes.