Implement GetSchemaData and PutSchemaData for the disk backend.
This means that all the backends now implement these functions, so they
can go into the ISchemaBackend interface.

They also now need to take a sequence numbers, as the data files need to
live on disk in a particular order.  The db and dbus backends ignore this
number for now.