Evil hack to get the schema of a particular request added to WvDbi.
WvDbi is pretty amazing in every conceivable way, except for the issue
of wanting to get the schema back from a set of results.  While
WvSqlRow largely fixed this problem, it doesn't do it if you perform
an SQL "SELECT" and you don't get back any rows.

Hence; WvDbi.statement_schema added, which just performs a query and
returns the results of the resulting IDataReader.GetSchemaTable().