Split up schemamatic.t.cs.
Having a 1300-line unit test wasn't very sane, and it was slow to run.  As
a first cut, put all the VxDiskSchema tests in their own test file, and
move some of the shared utility functions into their own class, shared
between the original schemamatic.t.cs and the new vxdiskschema.t.cs.

This could definitely still stand to be split up a bit more, but it's a start.