Turn WvAutoCasts of booleans into integers.
Previously, casting a WvAutoCasted boolean to a string would get you
"True" or "False", and casting it to an int would get you just plain 0.
It makes more sense to get "1" and "0" out in string form, and it
makes a lot more sense to be able to get both 1 and 0 out in int form.