Use "SET IDENTITY INSERT ON/OFF" statements when needed.
When inserting data into the database, if you want to explicitly set
the values of primary key columns, you have to say "SET IDENTITY INSERT
ON" before you start inserting.

The Perl code had the advantage of already knowing whether a table
had a primary key column or not when it was creating its insert
statements.  We now have to go and ask the database again about each
table.