versaplexd now always connects using WvDbi.
It never needs to create its own SqlConnection anymore, although it
occasionally still accesses the inner fixme_db member (for no good reason
except that I'm too lazy to fix it).  Similarly with the unit tests:
versaplextester.cs is pretty revolting, but no time to clean it up right
now.

Anyway, this is much cleaner than it was.