Added wvdbi bookmarks and made wvini writable.
WvDbi bookmarks now auto-save themselves to ~/.wvdbi.ini.  When you connect
to a database with a long URL, it'll try to save that URL under a convenient
name (generally the database name) in wvdbi.ini automatically for easier
typing later.

This required finally adding write support to WvIni.  While I was there, I
also switched it to use new-style Dictionary<K,V> instead of the old
Hashtable class, which required widespread changes to make it work properly
again, for better or for worse.