Don't pretend dbus-sharp is built if dbus-sharp/Makefile doesn't exist.
Previous versions had a built-in Makefile, but now it's generated by
autoconf.  Watch out for that, and prevent people from getting confused when
a toplevel 'make' doesn't build everything.  (./configure was running
autoconf in the dbus-sharp directory, but it's easy to forget to do that
when checking out a new branch.)