wvdbi: create a new WvSqlRows object that holds a schema *and* an enumerable.
We're working here toward eliminating the horrible WvDbi.statement_schema()
hack job, which is evil on multiple levels (not least of which is that it
results in your query running twice... sometimes!).  The problem is that a
zero-row resultset won't have anything in the IEnumerable<WvSqlRow>, and
only the row members have a pointer to the schema, so you can't get at the
schema.  Of course, since there are no rows anwyay, you might think you
don't care.  Well, apparently that's not the case.

So we introduce a new WvSqlRows object that can act as a list of rows *and*
has a columns member like each row does.  So now you can do:

	using (var resultset = dbi.select("stuff"))
	{
		// do stuff with resultset.columns

		foreach (var r in resultset)
		{
		}
	}

However, the outer using() is optional if you don't care about the schema,
or if you only care about the schema when there's at least one row.

We don't actually remove statement_schema() yet: that's next.