Fix failing versaplexd VerifyDecimal unit test by hacking up WvData.
Somehow I missed that this unit test started failing after the recent
changes to use WvDbi in the unit tests.  It's because in mono,
IDataReader.GetValues() unfortunately explodes when you have crazy
decimal(38,38) columns, and you have to use SqlDataReader.GetSqlDecimal()
instead on the columns in question.  I don't even want to know why that
would be.  Anyway, we now do that automatically in WvData in case of an
explosion, which unfortunately makes wvdata.cs require
System.Data.SqlClient (ie. a client-specific library), which it didn't
before.  Oh well, the rest of wvdotnet requires it anyway at the moment.