Improved vxcli to (partly) work over DBus to a versaplexd.
You can now use a moniker of "vx:" to connect to the versaplexd on the
DBUS_SESSION_BUS_ADDRESS and issue queries.  Unfortunately, when you get an
actual resultset back, we only parse the column information, not the data
table itself, because it seems like parsing variants in dbus-sharp might be
a giant pain.  Sigh.