ExecChunkRecordset support in VxODBC.
In order to allow for ExecChunkRecordset, VxODBC assigns a callback
handler to all WvDBusConns, which processes signals containing data.

A map contains a list of outgoing (request) serial #'s mapped to
VxResultSet *'s, and this map is checked by the callback; if an
incoming signal is destined as a response to a particular serial #
in said map, the corresponding VxResultSet processes the signal.