Modified ExecChunkRecordset to use constant signal names.
Previously, Versaplexd would return a signal suffixed with the serial
of the query this was a response to.  This isn't really very nice,
since it means that the signals have different names depending on the
serial number we sent out.  Fudge that.

Instead, we now suffix a uint32 to the end of a signal response, which
can quickly be picked out for message processing, and current handlers
ignore it (data past the end of what they're expecting).