Store tables in a parseable format instead of CREATE TABLE statements.
Storing tables as CREATE TABLE ... statements means that the only way
to update a table is to drop it and re-add it.  Instead, store tables
as a more parseable format, with one line per table column.  This also
lets us store the table's indexes (including primary keys) in the
same place as the table itself.

For now, the index checksums are still returned separately, and tables
are still modified by dropping and re-adding them.