Name the primary keys by default.
If we don't name the primary key of a table, MS SQL will assign it an
annoying pseudo-random name.  Instead, name the primary key of a table
PK_TableName by default.  If you don't like that, you can specify a
name for the primary key by giving a name=MyName parameter on the
primary-key: line.