vxsanity.t.cs now passes if run from Mono against versaplexd on MS .NET
This required some hacking to do, as the way that vxsanity checks for
sanity is to make a versaplex query, while also independently
connecting to the database and making its own query.  Great idea,
but if we're running Mono and Versaplexd isn't, you get some
interesting inconsistencies.  This test is now updated to account
for those inconsistencies if they come up.

Unfortunately, it's also *ever so slightly* less thorough than
before, as some tests are skipped based on checking just a few
values.  Perhaps as future versions of Mono come out, they can be
put back in.