Don't assume that the database starts out empty in the unit tests.
Sometimes there'll be stray database elements around when we start the
unit tests.  Try to account for that possibility, and don't fail when
they show up.  Generally speaking, this involves counting the number
of elements that are around before starting, and testing that
that number only changes by the expected amount.