Versaplexd benchmarking utility!
OK, it's a bit of a pain to set up, but that's not my fault.  Check out
the header of the file for all the set-up instructions and variables to
tweak.  Then, knock yourself out trying out various combinations of
Versaplexd + dbus + protocol + network to see what you can see!