Sort table checksums.
MSSQL doesn't sensibly let you add new columns in the middle of a table.
So if you add a new column, it'll get inserted at the end of the table.
But then if you recreate the table from scratch, the column will end up
in the right spot, but your checksums won't match up.  Sorting the table
checksums means that (basically) equivalent tables will still have
matching checksums; just don't rely on the column order of the table.

Also move VxSchema.ParseKey to VxSchemaChecksums, as it now needs to
parse keys, but doesn't want to link with all the stuff that VxSchema
links to.  Leave a utility function in VxSchema to forward, as that
ends up making more sense API-wise.

Also rename VxSchemaChecksum's "name" field to "key", to match the
naming conventions used everywhere else.