Only change the parts of tables that need changing.
Dropping and adding a table just to tweak a column's type is obviously
annoying.  Now that tables are stored in a more sensible format, we
can figure out exactly what needs to change when a table is updated, and
only do that.  This means that the only changes that need to set the
Destructive parameter to Put are ones that drop a column.

Also returns the name of the primary key if it differs from the default
"PK_TableName" value.