Add some error handling when modifying a table, and also a transaction.
If an error occurs when modifying a table, roll back the transaction to
before we started changing anything.  Otherwise we could leave the table
in a bit of a mess.

Also move the updating of a single table to its own function, which
makes the above a lot easier to accomplish, and generally improves
sanity anyway.