Allow renaming indexes and primary keys.
There's no point in ever trying to change indexes or primary keys,
they're always safe to drop and re-add.  We weren't actually trying to
change them, but the function that computed a table diff was still
marking them as "Changed" entries.  Instead, put them in the diff as a
delete and an add, which means that the code doing the deleting and
adding has access to both the old and new names in the case of a name
change, which means primary key and index renames now work.  Yay!