Add a few features to WvData to be more IronPython-friendly.
WvAutoCast is great for C#, but not so great for python, which is happier
with a plain array of Object: it knows how to peer into the array and figure
out whether an object is an int, a string, etc.  So add a way to get the
object[] array back out.

While we're there, fix up the idiotic DBNull objects being inserted into our
object[] array by System.Data.  I want null, not an object that claims to be
null, thanks.

The bad news is that we still have a non-null WvAutoCast even if the entry
itself is null, but at least that's for a reason: the WvAutoCast has to
exist so we can cast it to things.