Monikerized VxWhateverSchema; now sm.exe doesn't require versaplexd.
The second parameter to sm is now a VxSchema moniker.  The monikers have a
few special rules for extra convenience: if the name matches a directory, we
prepend dir: automatically.  If the moniker can't be found, we prepend dbi:
automatically.

So, for example:
	sm pull mssql://server/db schemadir
	sm pull schemadir schemadir2
	sm push vx: schemadir2

The second parameter isn't a moniker yet, because I'm not sure GetSchemaData
and PutSchemaData would be able to deal with it.