Drop indexes outside the transaction, and roll them back by hand.
If you drop an index inside a transaction in MSSQL, you still can't modify
any columns covered by that index until you commit the transaction.  This is
a pain for us.  Instead, drop all indexes outside the transaction, and if
we have to roll back, reapply the indexes by hand.