Use StringBuilder instead of WvBuf to build the string in hexdump.
StringBuilder is much, much faster than WvBuf when you're adding a lot of
small strings.  With this change, I can now hexdump 100 MB of data in the
time it used to take for 1 MB.

Also don't call wv.fmt unnecessarily, when it was just used to tack on a
space.