Add and change columns in the same order they're given.
Before, we were sorting the lists of columns to add and change alphabetically
by name, which meant that the schema you asked for might not have been the
schema you got, even when it was perfectly possible to do it right.  It's
not really much extra work to keep the list of things to change in the
right order.