Remove some special case code in VxSchemaTable.GetDiff.
Since tables now have a dictionary of table elements, we don't have to do
anything special about changing indexes or primary keys in
VxSchemaTable.GetDiff; it's just as easy for the caller to do it now, if
they want to.  The only reason we were doing it before was it was hard
to dig out the old schema element once GetDiff had returned.