Added a registry writing application for SSL certs to VxODBC.
In order to support SSL, VxODBC needs a few values written into the
registry so that it would know who to connect to, and what certificate
to present.  Added 'regwriter.exe,' a tool that takes a .ini file
specifying your certificate, private RSA key, and the certificate of
the DBus server you expect VxODBC to connect to, and writes them
(encrypted via CryptProtectData) the Windows registry, where VxODBC
will find them when it runs.  See the regwriter.cc file for .ini
file format, and also the vxodbc.ini.tmpl file for an example.