setup_starquery_bench.pl; to test VxODBC/MSSQL StarQuery performance.
Run this script to populate an MSSQL database with a bunch of tables,
which can be used with a StarQuery map file (not included in this
commit, since the map files have to be defined for a specific
connection, driver, etc. and anyone with different ones than me can't
benefit from the work anyway).  I used this to test VxODBC vs.
MSSQL TDS performance.