Support adding and removing default constraints on columns.
Default constraints can't be changed in an ALTER TABLE ... ALTER COLUMN
statement, they're treated as separate constraint objects.  So drop and
add them explicitly.  This means that they can now be changed without
specifying the Destructive option to Put.

While I was doing this, I noticed that my transaction handling was rather
broken.  Any rollback will roll back all outstanding transactions, and most
of the errors we have to deal with will automatically (and non-optionally)
do a rollback.  So to recover from errors that might occur when we alter
a column, we have to do all the column altering ahead of time in a
completely separate transaction, remember any error messages, and then roll
back.