Rename VxSqlException.GetFirstSqlErrno() to Number.
It turns out that plain SqlExceptions have a Number property that does
the same thing that GetFirstSqlErrno() does.  Rename GetFirstSqlErrno() to
match, and then use the SqlException's Number property to implement it.

Also get rid of a few spots that were creating a VxSqlException just to get
easy access to the first error number.