If adding or changing a column fails due to being not null, make it null.
It's more important to be able to apply a schema change than to keep data,
at least if the Destructive flag is set.  So if a column addition or update
fails because it's adding a non-null column to a table that already has
data, we should just make the column be nullable if the Destructive flag
is set.  You may lose a bit of data, but, well, you asked for it.