Drop non-null constraints if adding a column in non-destructive mode too.
If we're in non-destructive mode, and adding a new non-null column to a
table with data in it, it's fine to drop the nullability constraint (since
we'll send back a warning).  It doesn't destroy any data, so we don't have
to require the Destructive flag.

Also, don't send the "non-null column contains nulls" warning if we had
any other errors, they may just confuse things, and send a different
warning if we run into problems figuring out if there are nulls.