If updating a table in destructive mode has an error, delete and re-add it.
Destructive mode means we're allowed to delete what we like in order to
make the database like what was presented.  So we shouldn't settle for
"almost right" schemas.  If anything goes wrong, stop playing nice and
drop and re-add the table.

Also check explicitly for trying to create a table with no columns in
it, as the error message is otherwise not very enlightening, and let
VxSchemaError take a generic VxRequestException instead of just a
VxSqlException.