Drop null constraints when modifying a column, in addition to adding columns.
If you pushed a schema with a new non-null column to a table with data,
Schemamatic would drop the null constraint (if you didn't specify
Destructive) and just give you a warning.  But if you then pushed the
exact same schema again, it'd fail, because we didn't know how to drop
the null constraints when modifying a column.

Anyway, now we can drop null constraints when modifying a column too,
so you can push the same schema repeatedly and get the same results.