vxodbc/t/vxodbctester.{cc,h}:  Fixed total evil static variable.
Removed a static variable and made it non-static.  The evil this
variable was causing is too much for me.  It must have been causing
timing errors since the dawn of time, since it persists across
multiple VxOdbcTester objects.  I don't know how the code ever
worked with it.  The world is a better place now that it's gone.