versaplexd no longer needs to use the session bus for checking userids.
Now that the default DBus transport is sane, we can use it for everything.
Of course, it was a little *too* sane and was happily processing requests
recursively, so I had to linearize the requests using a Queue<Action> in
versaplexd.cs.  Otherwise we might handle incoming database query requests
while we waited for replies needed to handle the previous request, and we'd
end up running queries out of order.