Establishing a baseline of errors that occur is now useless.
It used to be that if you called Get and Put on a table, without
specifying the Destructive option, it'd give you an error.  Putting an
unchanged table now doesn't do anything, so doesn't give an error,
so checking for baseline errors will never give you anything useful.
And if it ever started, there's a test right below that checks that
Get and Put don't ever cause errors.  Oops.