versaplexd:  First commit of an SQL Tokenizer.
Our woes with parsing queries should be alleviated with this handy-
dandy T-Sql tokenizer.

I've conveniently converted vxapi.cs to use it in the instances
where we have to parse and transform queries.  Passes all unit
tests now, but could probably use some improvement in the
interface, which is a bit messy.