Use VxSchemaElement.create instead of a public constructor.
There were a bunch of special cases - and at least one too few of them -
that created a new VxSchemaTable for elements of type Table, else a new
VxSchemaElement.  Now we have a static member in VxSchemaElement that will
do that for you.

This fixes a bug where "sm push" wasn't able to push new tables from a file
to a database, instead interpreting the table schema information as valid
SQL and sending it directly.