More unit test fixes: can't create a corrupt VxSchemaTable.
If VxSchemaElement.create() tries to create a VxSchemaTable but it throws an
exception, it now catches it and creates a plain VxSchemaElement instead.

This is a little fishy, but it's better than the alternatives, since we have
to go through here even if we're just checking checksums.  And we should
still be able to get a checksum of the data even if the data is unparseable.