vxodbc/vxhelpers.*: Add function to list '__VERSAPLEX' as only catalog.
When a calling application wishes to know what catalogs are available
to vxodbc on our Versaplex server, we plan on returning '__VERSAPLEX',
and masking what's really going on.  To this end, add the
'return_versaplex_db()' function, which properly (with column/row
info) simulates the return of a query to the server listing available
catalogs.