vxodbc/connection.cc:  Modify CurrCat to return "__VERSAPLEX" catalog.
As part of our tactic to get StarQuery to list "__VERSAPLEX" as our one
and only catalog, it appears that SQLGetInfo has to return a non-empty
value for our current catalog when called with the SQL_DATABASE_NAME
InfoType.  Very well.  Fake it by returning "__VERSAPLEX" for non-
MS Query clients, and a null string if we do use MS Query.

Micro-optimisation in CurrCatString as well.