vxodbc/info.cc:  The fix to the "initial catalog error" in StarQuery.
Now that all the pieces are in place, StarQuery calls SQLGetInfo and
gets back a non-empty value, it's going to call SQLTables with wildcard
parameters to try to get a list of our catalogs.  Rather than ever
contacting Versaplex, we'll just fake a response by calling the
return_versaplex_db function from a few commits back!