Finally!  StarQuery reports dollar signs on decimal fields.
This required setting all fields of type Decimal to have pgtype_money=true,
as well as precisely setting the column size/decimal digits to 19 and 4,
respectively.  In other words, it's kind of hacky.  But at least it's
basically working.

To make it less hacky, we'd have to get (and keep) the
size/precision/moneyness data from the remote end somehow, which I leave as
an exercise to the reader.  (It can't very very hard, since the we're
talking to the MSSQL ODBC driver, which clearly provides that information.)