Add a unit test for user-defined data type problem.
Also fixed some minor test problems: include a "drop procedure Func2" as
well as "drop function Func2", since some tests apparently create it as a
procedure and might accidentally leave it lying around.

Also, when we run SQL statements in our tests and we expect they might fail,
you can now use VxExecSilent instead of catching the exception yourself.  It
politely prints out the error and then continues, which can be helpful with
debugging (which it was for the above bug).