versaplexd/{versaplexd.ini.tmpl,vxsqlpool.cs}:  New security settings.
Added support for limiting the amount of DB-modifying queries a user
can issue in a single Versaplex query to 1, if the security settings are
so set.  Please read the comments in the new versaplexd.ini.tmpl file for
more information.

NOTE:  If you're currently testing versaplex, you should *really* add the
[Security Level]
* = 1

lines to your versaplexd.ini.