versaplexd/vxapi.cs:  Based on security level, limiting of #commands.
Added code to vxapi.cs which iterates through a query and determines
if it's 'secure' based on our new security options.

I find this code ugly and hacky myself.  The proper solution is to
create an SQL parser, not this trial-and-error hacky exceptions for
various commands code.

Regardless, we can now *sort of* limit the ability of anyone to slip in
too many non-select statements in a single query (and yet not cripple
people's ability to use the database).

The code throws a VxSqlException telling the user that he was being
sneaky if it catches a user with security turned on him attempting to
send through multiple non-select statements.