Fix (well... improve) race condition in cancelquery.t.cs.
A "select 1" statement is just way too fast sometimes, apparently, and the
server doesn't get the cancel request before it finishes the select
statements.

Now we delay for 1 second instead, so we have a pretty high probability that
it'll get there... but it's still an arbitrary delay, which is kind of
annoying.