versaplexd/vxapi.cs:  Retry a connection to MSSQL if it fails.
Due to the silly way in which connections to restarting MSSQL servers
happen, we should retry a failed connection attempt once to try to get
in touch with our SQL roots.

This code implements this (only in ExecChunkRecordset, because that's
really the only command we call outside of the unit tests, and the only
command VxODBC calls) by handling a connection error exception, and then
re-entering a while-loop around the connection code.

If another exception is thrown, we propagate it up and exit the loop,
as normal.