versaplexd-svc: add description field, logfile, and catch exceptions.
We weren't setting the "Description" field in the Services control panel.

Also, we should search for versaplexd.ini in the same directory as the dll,
which is apparently *not* the default directory the service runs from (on
Windows).

Use WvLog's new log-to-file support to write to a file called versaplex.log,
also in the same directory as the exe.  Writing to stdout, of course, makes
little sense in a Windows service.

We also do a little work toward catching exceptions.  We log exceptions that
occur in Versaplex.Go instead of letting them leak all the way to the top
level, where they would be unnecessarily scary (and you don't get to read
the actual exception message, sigh).

We now don't let you install the Windows service at all if we can't find the
config file.