versaplexd-svc: add -n (service name) option.
You might want to install more than one versaplexd on the same machine, I
suppose.  Or in my case, stupid Windows "sometimes" doesn't finish removing
a service until you reboot, so to experiment, you have to keep renaming the
service to a new name each time you do a test install.  Morons.

Also, don't rollback the install just because we failed to start the
service.  That would make it impossible for the admin to look into why the
service failed to start in the first place (and then fix it).  The old
behaviour was just an artifact of putting the statement inside a try{} block
by accident.