Make schemamatic fucking fast.
OK, I'll be honest, I'm not *entirely* sure what caused this.  Was it
the ripping out of WvCsv's guts, so that it iterates over a
WvInBufStream, rather than taking a bunch of text?  Was it the calling
of GetColType only once per column for a particular table, rather than
n * m times where n is the number of rows, and m is the number of
columns?  Was it the formatting of the INSERT statement only once,
rather than once per row?

This giant changeset produces a sqlite.db identical to our reference
sqlite, except in 38 seconds instead of 9 minutes.