Removed wvcsv.cs.
One day, I'm sure there'll be a need for it.  Right now, it's a
worthless class, it's read-only, and keeps no state information.  As
such, it's just not worth having it sit there individually; it's
easily broken down into two functions which might as well sit in
vxdbschema.cs, since it's WvCsv's only caller.

Runtime for OTBET reduced to 32 seconds.