Ugly hack to speed up dpull
In order to be able to pull large amount of data (e.g. more than the
available RAM) we must not use sorting because that implies we'd have
to hold ALL data in RAM to be sorted later. For really large tables we
end up out of memory in a crash. New command export-nosort added.